- Confirmation box before deleting the page.

+++ src/components/DynamicPageNav.js
@@ -2,8 +2,13 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
+import UIkit from 'uikit';
 import DynamicPageMenuItem from './DynamicPageMenuItem';
 import { deletePage } from '../redux/page/pageActions';
+
+const EDIT_PAGE_ID = 'edit-page-id';
+const DELETE_PAGE_ID = 'delete-page-id';
+const CREATE_SUB_PAGE_PAGE_ID = 'create-sub-page-page-id';
 
 class DynamicPageNav extends Component {
   onEditPage = () => {
@@ -12,9 +17,15 @@ class DynamicPageNav extends Component {
   };
 
   onDeletePage = () => {
-    const { deletePageDispatch, bookId, pageId, history } = this.props;
-    deletePageDispatch(bookId, pageId);
-    history.push('/');
+    UIkit.util.on(`#${DELETE_PAGE_ID}`, 'click', (e) => {
+      e.preventDefault();
+      e.target.blur();
+      UIkit.modal.confirm('Do you want to delete this page?').then(() => {
+        const { deletePageDispatch, bookId, pageId, history } = this.props;
+        deletePageDispatch(bookId, pageId);
+        history.push('/');
+      });
+    });
   };
 
   onCreateSubPage = () => {};
@@ -44,18 +55,21 @@ class DynamicPageNav extends Component {
           icon="file-edit"
           text="Edit this page"
           onClick={this.onEditPage}
+          id={EDIT_PAGE_ID}
         />
         <DynamicPageMenuItem
           key="trash"
           icon="trash"
           text="Delete this page"
           onClick={this.onDeletePage}
+          id={DELETE_PAGE_ID}
         />
         <DynamicPageMenuItem
           key="new"
           icon="plus"
           text="Create new sub page"
           onClick={this.onCreateSubPage}
+          id={CREATE_SUB_PAGE_PAGE_ID}
         />
       </div>
     );
@@ -80,7 +94,7 @@ DynamicPageNav.propTypes = {
 
 const mapStateToProps = ({
   nav: {
-    active: { book, page },
+    active: { book = '', page = '' },
   },
 }) => {
   return {
